+++ spec/ig_markets/transaction_spec.rb
@@ -1,4 +1,4 @@
-describe IGMarkets::DealingPlatform do
+describe IGMarkets::Transaction do
   it 'correctly identifies interest payments' do
     {
       { transaction_type: :with, instrument_name: 'interest' } => true,
@@ -8,15 +8,15 @@ describe IGMarkets::DealingPlatform do
       { transaction_type: :with, instrument_name: 'pinterest' } => false,
       { transaction_type: :depo, instrument_name: 'interests' } => false
     }.each do |attributes, result|
-      expect(build(:account_transaction, attributes).interest?).to eq(result)
+      expect(build(:transaction, attributes).interest?).to eq(result)
     end
   end
 
   it 'reports correct profit/loss amounts' do
-    expect(build(:account_transaction, currency: '$', profit_and_loss: '$1,000.00').profit_and_loss_amount).to eq(1000)
-    expect(build(:account_transaction, currency: 'AU', profit_and_loss: 'AU-5.50').profit_and_loss_amount).to eq(-5.5)
+    expect(build(:transaction, currency: '$', profit_and_loss: '$1,000.00').profit_and_loss_amount).to eq(1000)
+    expect(build(:transaction, currency: 'AU', profit_and_loss: 'AU-5.50').profit_and_loss_amount).to eq(-5.5)
 
-    expect { build(:account_transaction, currency: 'USD', profit_and_loss: 'EUR-5.0').profit_and_loss_amount }
+    expect { build(:transaction, currency: 'USD', profit_and_loss: 'EUR-5.0').profit_and_loss_amount }
       .to raise_error(StandardError)
   end
 end